dj_estacionamiento/requirements.txt
Django==5.0.2
pillow==10.2.0
djangorestframework==3.14.0
drf-extra-fields==3.7.0
opencv-python==4.9.0.80
pytesseract==0.3.10